Move existing GUI state from Preferences to GuiSettings
